fix: correct property type for cachedAvailableCores to int
SystemMetrics::limits()->availableCpuCores() returns int, not float.
Correcting the property type eliminates a PHPStan property.unusedType error.

+++ src/Scaling/Calculators/CapacityCalculator.php
@@ -22,7 +22,7 @@ class CapacityCalculator
 
     private ?float $cachedTotalMemoryMb = null;
 
-    private ?float $cachedAvailableCores = null;
+    private ?int $cachedAvailableCores = null;
 
     private ?float $cacheTimestamp = null;
